add eaforum-specific cors response headers to the /graphql route to fix crosspost account linking
