+++ app/api/v2/crosspost/connectCrossposter/route.ts
@@ -1,4 +1,5 @@
-import { connectCrossposterCrosspostHandler, crosspostOptionsHandler } from "@/server/crossposting/handlers";
+import { connectCrossposterCrosspostHandler } from "@/server/crossposting/handlers";
+import { crosspostOptionsHandler } from "@/server/crossposting/cors";
 import type { NextRequest } from "next/server";
 
 export function POST(req: NextRequest) {
