+++ app/api/v2/crosspost/createCrosspost/route.ts
@@ -1,4 +1,5 @@
-import { createCrosspostCrosspostHandler, crosspostOptionsHandler } from "@/server/crossposting/handlers";
+import { createCrosspostCrosspostHandler } from "@/server/crossposting/handlers";
+import { crosspostOptionsHandler } from "@/server/crossposting/cors";
 import type { NextRequest } from "next/server";
 
 export function POST(req: NextRequest) {
