+++ app/api/v2/crosspost/unlinkCrossposter/route.ts
@@ -1,4 +1,5 @@
-import { crosspostOptionsHandler, unlinkCrossposterCrosspostHandler } from "@/server/crossposting/handlers";
+import { unlinkCrossposterCrosspostHandler } from "@/server/crossposting/handlers";
+import { crosspostOptionsHandler } from "@/server/crossposting/cors";
 import type { NextRequest } from "next/server";
 
 export function POST(req: NextRequest) {
